feat(admin): refactor analytics and video management logic with Redis integration
