+++ apps/admin/package.json
@@ -6,6 +6,7 @@
     "@sentry/nextjs": "10.20.0",
     "@sentry/profiling-node": "10.20.0",
     "@shinju-date/constants": "workspace:*",
+    "@shinju-date/helpers": "workspace:*",
     "@shinju-date/temporal-fns": "workspace:*",
     "@supabase/sentry-js-integration": "0.3.0",
     "@supabase/supabase-js": "2.75.1",
